Move store types to own file
