+++ packages/client-web/src/store/index.ts
@@ -19,8 +19,4 @@ const store = createStore(
   composeWithDevTools(applyMiddleware(thunkMiddleware)),
 );
 
-export type GetState = typeof store.getState;
-export type RootState = ReturnType<GetState>;
-export type AppDispatch = typeof store.dispatch;
-
 export default store;
