+++ packages/client-web/src/store/types.ts
@@ -1,0 +1,5 @@
+import store from ".";
+
+export type GetState = typeof store.getState;
+export type RootState = ReturnType<GetState>;
+export type AppDispatch = typeof store.dispatch;
